+++ __tests__/Car.test.js
@@ -1,6 +1,6 @@
 import Car from "../src/domain/Car.js";
 
-import { getCarStatus } from "../src/util/carUtil.js";
+import { getCarStatus } from "../src/service/carService.js";
 describe("차 객체 유닛 테스트", () => {
   test("차 객체를 만들면 차가 생성되어야 함", () => {
     //given
